requirements.txt
django-json-field
